authd: fix auth->cid type sizes
* long is too small on 32-bit systems, use unsigned long long if we want
  to check for out of range values
* UINT32_MAX is a valid cid, and 0 isn't
* make auth->cid a uint32_t not uint16_t

+++ authd/provider.h
@@ -44,7 +44,7 @@ struct auth_client_data
 
 struct auth_client
 {
-	uint16_t cid;				/* Client ID */
+	uint32_t cid;				/* Client ID */
 
 	char l_ip[HOSTIPLEN + 1];		/* Listener IP address */
 	uint16_t l_port;			/* Listener port */
